Sprint 7

+++ local.dbaid.database/Database/Table/dbo_config_job.sql
@@ -8,7 +8,7 @@ CREATE TABLE [dbo].[config_job]
 (
 	[job_id] UNIQUEIDENTIFIER NOT NULL PRIMARY KEY,
 	[job_name] NVARCHAR(128) NOT NULL,
-	[max_exec_time_min] SMALLINT NOT NULL,
+	[max_exec_time_min] INT NOT NULL,
     [change_state_alert] NVARCHAR(8) NOT NULL DEFAULT N'WARNING',
 	[is_enabled] BIT NOT NULL DEFAULT 1,
 	CONSTRAINT [CK_config_job_state] CHECK ([change_state_alert] = N'NA' OR [change_state_alert] = N'OK' OR [change_state_alert] = N'WARNING' OR [change_state_alert] = N'CRITICAL')
